refactor: remove uncessary file and make schema as private property
